Files verplaatst naar folder + suggestie van ander team om datetimes / users vast te leggen. Todo; nog verwerken in keys
