--- ms18-database/ReceiptStatus.cs
@@ -1,15 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Maasgroep.Database
-{
-	public record ReceiptStatus
-	{
-		[Key]
-		public short Id { get; set; }
-		public string Name { get; set; }
-
-
-		//Ef
-		public ICollection<Receipt> Receipt { get; set; }
-	}
-}
