+++ ms18-database/teamC/Photo.cs
